requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.1
drf-spectacular>=0.27.2
psycopg2-binary>=2.9
pdfplumber>=0.11.4
pytesseract>=0.3.10
PyPDF2>=3.0.1
python-dotenv>=1.0.1
gunicorn>=21.2.0